requirements.txt
langchain
langchain-community
langchain-openai
openai
streamlit
gradio
tiktoken
python-dotenv
pinecone-client
duckduckgo_search
python-dotenv
google-search-results